+++ inc_helper.rb
@@ -1,4 +1,5 @@
-require_relative 'add_item_info'
+require_relative 'module/add_item_info'
+require_relative 'module/save_load_book'
 require 'date'
 require 'json'
 require_relative 'item_class'
@@ -9,5 +10,3 @@ require_relative 'book'
 require_relative 'label'
 
 require_relative 'music_album'
-require_relative 'genres'
-
