+++ labels.json
@@ -1,7 +1,17 @@
 [
   {
-    "id": 5791,
+    "id": 1018,
     "title": "Kathy",
     "color": "pink"
+  },
+  {
+    "id": 6060,
+    "title": "John and Ryuuji",
+    "color": "green"
+  },
+  {
+    "id": 5903,
+    "title": "john",
+    "color": "blue"
   }
 ]
