+++ module/add_item_info.rb
